requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
streamlit>=1.35.0
python-dotenv>=1.0.0
